requirements.txt
psycopg2-binary>=2.9.0
Pillow>=10.0.0
transformers>=4.30.0
torch>=2.0.0
